splitting the engine into modules
