+++ modules/camera/headers/camera_s.hpp
