+++ modules/input/implementation/glfwIO.cpp
@@ -1,4 +1,4 @@
-#include <headers/glfwIO.hpp>
+#include <input/headers/glfwIO.hpp>
 
 // process all input: query GLFW whether relevant keys are pressed/released this frame and react accordingly
 // ---------------------------------------------------------------------------------------------------------
