+++ modules/objectloader/headers/triangulated_mesh.hpp
