+++ modules/timing/headers/time.hpp
@@ -1,9 +1,11 @@
-#pragma once
+#ifndef TIME_H
+#define TIME_H
 
 // This header only exists so here are some inline vars for setting delta time
-
+// in future this header might get some implemmentations for a timer or something similar
 namespace deltatime{
     inline float first;
     inline float delta_time;
     inline float last;
 }
+#endif
