+++ modules/timing/implementation/time.cpp
@@ -1,0 +1,7 @@
+#include <timing/headers/time.hpp>
+
+/*
+    This is only a dummy CPP file
+    so git will ad this folder to the repo
+    I dont want to modify the git ignore for this case
+*/
